GEO-INFER-INTRA/requirements.txt
fastapi>=0.95.0
pydantic>=2.0.0
sqlalchemy>=2.0.0
elasticsearch>=8.0.0
rdflib>=6.0.0
mkdocs>=1.4.0
celery>=5.2.0
pyyaml>=6.0
jsonschema>=4.0.0
typer>=0.7.0
rich>=12.0.0
uvicorn>=0.20.0
requests>=2.28.0 